feat: add docker-compose.yml file to configure MariaDB service
The docker-compose.yml file is added to the project to configure the MariaDB service. This file specifies the version of Docker Compose to use and defines a service named "db" using the "mariadb" image. The service is set to restart always and maps port 3000 on the host to port 3000 in the container. It also sets the environment variable "MARIADB_ROOT_PASSWORD" to the value of the "MARIADB_PASSWORD" environment variable.